moved scraping tools to to the advanced scraper
